% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fft.R
\name{find_fft_peaks}
\alias{find_fft_peaks}
\title{Find the top-N local maxima in an FFT spectrum}
\usage{
find_fft_peaks(spectrum_df, n_peaks = 10L)
}
\arguments{
\item{spectrum_df}{data.frame returned by \code{compute_fft_spectrum}}

\item{n_peaks}{Maximum number of peaks to return}
}
\value{
data.frame with columns Frequency, Magnitude, Period
}
\description{
Find the top-N local maxima in an FFT spectrum
}
