% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_editor.R
\name{mod_editor_server}
\alias{mod_editor_server}
\title{Model code editor module — server}
\usage{
mod_editor_server(id)
}
\arguments{
\item{id}{Module namespace id}
}
\value{
Reactive string containing the current editor text
}
\description{
Model code editor module — server
}
