% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{mod_plot_server}
\alias{mod_plot_server}
\title{Plotly time-series plot module — server}
\usage{
mod_plot_server(id, result)
}
\arguments{
\item{id}{Module namespace id}

\item{result}{Reactive returning either a data.frame (success)
or an object of class "rhapsody_error" (failure)}
}
\description{
Plotly time-series plot module — server
}
