% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scan.R
\name{mod_scan_server}
\alias{mod_scan_server}
\title{Parameter scan module — server}
\usage{
mod_scan_server(id, ir, param_state)
}
\arguments{
\item{id}{Module namespace id}

\item{ir}{Reactive returning Canonical IR (or NULL)}

\item{param_state}{Return value of mod_params_server}
}
\value{
Invisibly NULL (called for its side effects — renders scan plot)
}
\description{
Parameter scan module — server
}
