% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_steadystate.R
\name{mod_steadystate_ui}
\alias{mod_steadystate_ui}
\title{Steady-state module — UI}
\usage{
mod_steadystate_ui(id)
}
\arguments{
\item{id}{Module namespace id}
}
\description{
Steady-state module — UI
}
