% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_syntax.R
\name{parse_model}
\alias{parse_model}
\title{Parse model source text into the Canonical Internal Representation}
\usage{
parse_model(source)
}
\arguments{
\item{source}{Character string of model code}
}
\value{
Named list with fields: meta, type, states, parameters,
auxiliary, events, time, raw_source
}
\description{
Parse model source text into the Canonical Internal Representation
}
