% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the rhapsody Shiny application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to shiny::shinyApp()}
}
\description{
Launch the rhapsody Shiny application
}
