% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_io.R
\name{session_from_json}
\alias{session_from_json}
\title{Load a session from a .rhy JSON file}
\usage{
session_from_json(path)
}
\arguments{
\item{path}{Path to the .rhy file}
}
\value{
Named list with fields: rhapsody_version, model, params, ics, solver,
or a rhapsody_error if the file is invalid
}
\description{
Load a session from a .rhy JSON file
}
