% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_solver.R
\name{mod_solver_server}
\alias{mod_solver_server}
\title{Solver settings module — server}
\usage{
mod_solver_server(id)
}
\arguments{
\item{id}{Module namespace id}
}
\value{
Reactive returning list(method, atol, rtol)
}
\description{
Solver settings module — server
}
